pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inventory-intelligence-platform"
version = "0.1.0"
description = "FastAPI service for tracking inventory risk, surfacing danger alerts, and sending notifications."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [{name = "SINDH"}]
dependencies = [
  "fastapi>=0.100.0",
  "uvicorn[standard]>=0.23.0",
  "sqlalchemy>=2.0",
  "pydantic-settings>=2.0",
  "openpyxl>=3.1",
  "scikit-learn>=1.4",
  "joblib>=1.3",
  "PyJWT>=2.8",
  "Jinja2>=3.1",
  "itsdangerous>=2.2",
]

[tool.setuptools]
py-modules = ["bootstrap"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
